requirements.txt
fastapi~=0.79.0
requests~=2.28.1
pydantic~=1.9.1
python-decouple==3.6
httpx~=0.23.0
uvicorn==0.16.0
fastapi-utils==0.2.1
fastapi-plugins==0.10.0
aioredis~=2.0.1
SQLAlchemy~=1.4.39
passlib==1.7.4
psycopg2==2.9.1;
python-jose==3.3.0